% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALL-data.R
\docType{data}
\name{ALL_colosimo}
\alias{ALL_colosimo}
\title{Acute Lymphoblastic Leukemia}
\format{
A data frame with 17 observations.
}
\usage{
ALL_colosimo
}
\description{
Survival times, in weeks, of 17 patients with acute leukemia. For these
patients, their white blood cell counts (WBC) were recorded at the time of
diagnosis. The variables are:
}
\details{
\itemize{
\item times: time-to-event (in weeks).
\item status: censorship indicators.
\item wbc: WBC, white blood cells counts.
\item lwbc: base-10 logarithms of the WBC.
}
}
\examples{
data(ALL_colosimo)
par(mfrow = c(1,2))
hist(ALL_colosimo$time, main="", xlab="Time (Weeks)")

}
\references{
\insertRef{Colosimo2006a}{EstimationTools}
}
\keyword{datasets}
