% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALL-data.R
\docType{data}
\name{ALL_colosimo_table_4_3}
\alias{ALL_colosimo_table_4_3}
\title{Acute Lymphoblastic Leukemia, table 4.3}
\format{
A data frame with 33 observations.
}
\usage{
ALL_colosimo_table_4_3
}
\description{
Survival times, in weeks, of 17 patients with acute leukemia. For these
patients, their white blood cell counts (WBC) were recorded at the time of
diagnosis. The variables are:
}
\details{
\itemize{
\item times: time-to-event (in weeks).
\item status: censorship indicators.
\item wbc: WBC, white blood cells counts.
\item lwbc: base-10 logarithms of the WBC.
\item Ag_plus: whether the subject expresses the Calla antigen
(\code{Ag_plus} = 1) or not (\code{Ag_plus} = 0).
}
}
\examples{
data(ALL_colosimo_table_4_3)
hist(ALL_colosimo_table_4_3$time, main="", xlab="Time (Weeks)")

}
\references{
\insertRef{Colosimo2006a}{EstimationTools}
}
\keyword{dataset}
