% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_fun.R
\encoding{UTF-8}
\name{detection_fun}
\alias{detection_fun}
\title{Detection functions in \code{maxlogL} framework}
\usage{
detection_fun(fun, support, routine = "integrate", ...)
}
\arguments{
\item{fun}{a detection function. It must exist in Global Environment or loaded
from a package.}

\item{support}{a list with the following entries:
\itemize{
\item \code{interval}: a two dimensional atomic vector indicating the
set of possible values of a random variable having the
distribution specified in \code{y_dist}.
\item \code{type}: character indicating if distribution has a
\code{discrete} or a \code{continous} random variable.
}}

\item{routine}{a character specifying the integration routine.
\code{integrate} and \code{gauss_quad} are available, but the
custom routines can be defined.}

\item{...}{further arguments passed to the integration routine.}
}
\value{
A function with the folling input arguments:
\item{x}{vector of (non-negative) quantiles.}
\item{...}{Arguments of the probability density/mass function.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function takes the name of a key function and computes the corresponding
detection function.
}
\examples{
library(EstimationTools)

#--------------------------------------------------------------------------------
# Example 1: Half normal detection function without adjustment term.
g <- detection_fun(fun=half_norm, support=list(interval=c(0, 100),
                   type='continuous'))
curve(g(x, sigma=4.1058), to=100)

#--------------------------------------------------------------------------------

}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
