% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fibers}
\alias{Fibers}
\title{Tensile strengths}
\format{A data frame with 69 observations.}
\usage{
Fibers
}
\description{
Tensile strengths (in GPa) of 69 specimens of carbon fiber tested under tension
at gauge lengths of 20 mm.
}
\keyword{datasets}
