% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hazard_Shape.R
\name{Hazard_Shape}
\alias{Hazard_Shape}
\title{Summary of \code{HazardShape} objects}
\usage{
Hazard_Shape(object)
}
\arguments{
\item{object}{an object of class \code{HazardShape}, generated with
\code{\link{TTT_hazard_shape}}.}
}
\description{
This function displays the estimated hazard shape given a
data set.
}
\author{
Jaime Mosquera Gutiérrez \email{jmosquerag@unal.edu.co}
}
