% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrators.R
\encoding{UTF-8}
\name{integrate_stats}
\alias{integrate_stats}
\alias{gauss_quad}
\title{Integration using \code{integrate} from \strong{stats} package.}
\usage{
integrate_stats(fun, lower, upper, add_args = NULL, ...)

gauss_quad(
  fun,
  lower,
  upper,
  add_args = list(n = 10, kind = "legendre", normalized = FALSE, weighted = TRUE),
  ...
)
}
\arguments{
\item{fun}{an \strong{R} function which should take a numeric argument x and
possibly some parameters. The function returns a numerical vector
value for the given argument \code{x}.}

\item{lower}{a numeric value for the lower limit of the integral.}

\item{upper}{a numeric value for the upper limit of the integral.}

\item{add_args}{a list with further arguments for integration routine.}

\item{...}{additional arguments to be passed to \code{fun}.}
}
\value{
The value of the integral of the function specified in \code{fun}
argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function uses adaptive quadrature from \code{integrate} routine from
\strong{stats} package.
}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
\concept{Integration functions}
