% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{formula2Surv}
\alias{formula2Surv}
\alias{fo_and_data}
\title{Internal functions for formula and data handle}
\usage{
formula2Surv(model_frame)

fo_and_data(y, fo, model_frame, data, fo2Surv = TRUE)
}
\arguments{
\item{model_frame}{a model frame build internally in some functions on
\strong{EstimationTools} package}
}
\description{
Utility functions useful for passing data from functions
inside others.
}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
\keyword{internal}
