% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit_methods.R
\encoding{UTF-8}
\name{is.maxlogL}
\alias{is.maxlogL}
\alias{is.EmpiricalTTT}
\alias{is.HazardShape}
\title{Is return of any object of \code{EstimationTools}?}
\usage{
is.maxlogL(x)

is.EmpiricalTTT(x)

is.HazardShape(x)
}
\arguments{
\item{x}{Any object of \code{EstimationTools}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Checks if an object is any of the classes implemented in \code{EstimationTools}
package.
}
\author{
Jaime Mosquera Gutiérrez \email{jmosquerag@unal.edu.co}
}
