% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.options.R
\encoding{UTF-8}
\name{loess.options}
\alias{loess.options}
\title{Configure various aspects of LOESS in \code{TTT_hazard_shape}}
\usage{
loess.options(span = 2/3, ...)
}
\arguments{
\item{span}{the parameter which controls the degree of smoothing.}

\item{...}{further arguments passed to \code{\link[stats]{loess}} function.}
}
\description{
This function allows the user to set the parameters of \code{loess} function
used inside \code{TTT_hazard_shape}.
}
\details{
Please, visit \code{\link[stats]{loess}} to know further possible arguments.
The following arguments are not available for passing to the LOESS estimation:
\itemize{
\item{data}{The only data handled inside \code{TTT_hazard_shape} is the
computed empirical TTT.}
\item{subset}{This argument is used in \code{loess} to take a subset of data.
In this context, it is not necessary.}
}
}
\seealso{
\code{\link[stats]{loess}}, \code{\link{TTT_hazard_shape}}
}
\author{
Jaime Mosquera Gutiérrez \email{jmosquerag@unal.edu.co}
}
