% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlogLreg.R
\name{maxlogLreg}
\alias{maxlogLreg}
\title{Maximum Likelihood Estimation for parametric linear regression models}
\usage{
maxlogLreg(formulas, y_dist, data = sys.parent(), subset = NULL,
  fixed = NULL, link = NULL, start = NULL, lower = NULL,
  upper = NULL, optimizer = "nlminb", control = NULL, ...)
}
\arguments{
\item{formulas}{a list of formula objects. Each element must have an \code{~}, with the terms
on the right separated by \code{+} operators. The response variable on the left side is optional.
Linear predictor of each parameter must be specified with the name of the parameter folowed by
the sufix \code{'.fo'}. See 'Details' and the examples below for further
illustration.}

\item{y_dist}{a formula object tha specifies the distribution of the response variable.
On the left side of \code{~} must be the response, and in the right side
must be the name o de probability density/mass function. See the section
'Details' and the examples below for further illustration.}

\item{data}{an optional data frame containing the variables in the model. If data is not
specified, the variables are taken from the environment from which
\code{maxlogLreg} is called.}

\item{subset}{an optional vector specifying a subset of observations to be used
in the fitting process.}

\item{fixed}{a list with fixed/known parameters of distribution of interest. Fixed parameters
must be passed with its name and its value (known).}

\item{link}{a list with names of parameters to be linked, and names of the link object.
For names of parameters, please visit documentation of density/mass function.
There are three link functions available: \code{\link{log_link}},
\code{\link{logit_link}} and \code{\link{NegInv_link}}.}

\item{start}{a numeric vector with initial values for the parameters to be estimated.}

\item{lower}{a numeric vector with lower bounds, with the same lenght of
argument `start` (for box-constrained optimization).}

\item{upper}{a numeric vector with upper bounds, with the same lenght of
argument `start` (for box-constrained optimization).}

\item{optimizer}{a lenght-one character vector with the name of optimization routine.
\code{\link{nlminb}}, \code{\link{optim}} and
\code{\link[DEoptim]{DEoptim}} are available; \code{\link{nlminb}}
is the default routine.}

\item{control}{control parameters of the optimization routine. Please, visit documentation of selected
optimizer for further information.}

\item{...}{Further arguments to be supplied to the optimizer.}
}
\value{
A list with class \code{"maxlogL"} containing the following
 lists:
\item{fit}{A list with output information about estimation and method used.}
\item{inputs}{A list with all input arguments.}
\item{outputs}{A list with additional information:
      \itemize{
         \item Number of parameters.
         \item Sample size
         \item Standard error computation method.
         \item Number of regression parameters.
       }
}
}
\description{
Function to compute maximum likelihood estimators (MLE) of regression parameters
of any distribution implemented in \code{R} with covariates (linear predictors).
}
\details{
\code{maxlogLreg} calculates computationally the log-likelihood (log L) function
corresponding to the distribution specified in argument \code{y_dist} with linear
predictors specified in argument \code{formulas}. Then, it maximizes the log L through
\code{\link{optim}}, \code{\link{nlminb}} or \code{\link{DEoptim}}. \code{maxlogLreg}
generates an S3 object of class \code{maxlogLreg}.
}
\note{
The following generic functions can be used with a \code{maxlogL} object:
\code{summary, print}.
}
\examples{
library(EstimationTools)

#--------------------------------------------------------------------------------
# Estimation in simulated normal distribution
n <- 1000
x1 <- runif(n = n, -5, 6)
y <- rnorm(n = n, mean = -2 + 3 * x1, sd = exp(1 + 0.3* x1))

# It does not matter the order of the paramters
formulas <- list(sd.fo = ~ x1, mean.fo = ~ x1)

norm_mod <- maxlogLreg(formulas, y_dist = y ~ dnorm,
                       link = list(over = "sd", fun = "log_link"))
summary(norm_mod)
#--------------------------------------------------------------------------------
# Fitting with censorship
# (data from https://www.itl.nist.gov/div898/handbook/apr/section4/apr413.htm)

failures = c(55, 187, 216, 240, 244, 335, 361, 373, 375, 386)
fails <- c(failures, rep(500, 10))
status <- c(rep(1, length(failures)), rep(0, 10))
formulas <- list(scale.fo=~1, shape.fo=~1)

mod_weibull <- maxlogLreg(formulas, y_dist = Surv(fails, status) ~ dweibull,
                          start = c(scale=100, shape=10),
                          lower = c(scale=0,shape=0))
summary(mod_weibull)
#--------------------------------------------------------------------------------
}
\seealso{
Other maxlogL: \code{\link{maxlogL}}
}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
\concept{maxlogL}
