% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells-data.R
\docType{data}
\name{reduction_cells}
\alias{reduction_cells}
\title{Reduction cells}
\format{
A data frame with 20 observations.
}
\usage{
reduction_cells
}
\description{
Times to failure (in units of 1000 days) of 20 aluminum reduction cells.
}
\examples{
data(reduction_cells)
par(mfrow = c(1,2))
hist(reduction_cells$days, main="", xlab="Time (Days)")
plot(reduction_cells$days, xlab = "Cell (subjects)", lty = 3, type="h")

}
\references{
\insertRef{Whitmore1983}{EstimationTools}
}
\keyword{datasets}
