% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrators.R
\encoding{UTF-8}
\name{integrate_stats}
\alias{integrate_stats}
\alias{gauss_quad}
\title{Integration routines}
\usage{
integrate_stats(fun, lower, upper, add_args = NULL, ...)

gauss_quad(
  fun,
  lower,
  upper,
  add_args = list(n = 10, kind = "legendre", normalized = FALSE, weighted = TRUE),
  ...
)
}
\arguments{
\item{fun}{an \strong{R} function which should take a numeric argument x and
possibly some parameters. The function returns a numerical vector
value for the given argument \code{x}.}

\item{lower}{a numeric value for the lower limit of the integral.}

\item{upper}{a numeric value for the upper limit of the integral.}

\item{add_args}{a list with further arguments for integration routine.}

\item{...}{additional arguments to be passed to \code{fun}.}
}
\value{
The value of the integral of the function specified in \code{fun}
argument.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This family of functions use quadratures for solving integrals. The user can
create a custom integration routine, see \emph{details} for further
information.
}
\details{
\code{integrate_stats} uses \code{\link[stats]{integrate}} function from
\strong{stats} package, whereas \code{gauss_quad} uses the implementation of
gaussian quadratures from \strong{gaussquad} package.

The user can create custom optimization routines through implementation
of a wrapper function using the four arguments showed in emph{Usage} section.
}
\seealso{
\code{\link[gaussquad]{laguerre.quadrature}},
\code{\link[gaussquad]{legendre.quadrature}},
\code{\link[gaussquad]{chebyshev.c.quadrature}},
\code{\link[gaussquad]{gegenbauer.quadrature}}
}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
\concept{Integration functions}
